euclides da Cunha e Jonatas Emidio

+++ 2016/20160622 - Barbarossa - groovy/barbarossa.groovy
@@ -1,13 +1,13 @@
 def static armedSoldiers (soldiers, arms, daysToSnow) {
 
 	if (daysToSnow == 9) {
-		def armsPerSoldiers = arms/soldiers
-		def mod = arms%soldiers
+		def armsPerSoldiers = arms / soldiers
+		def mod = arms % soldiers
 
 		if (mod == 0) {
 			return armsPerSoldiers > 2 ? soldiers : 0
 		} else {
-			return armsPerSoldiers == 2 ? mod : 0
+			return armsPerSoldiers as int == 2 ? mod : 0
 		}
 	}
